calculator done

+++ src/main/java/pro/sky/java/course2/calculator/calculator/Service/CalculatorService.java
@@ -5,5 +5,5 @@ public interface CalculatorService {
     int calculateSum(int num1, int num2);
     int calculateDifference(int num1, int num2);
     int calculateMultiplication(int num1, int num2);
-    int calculateQuotient(int num1, int num2);
+    double calculateQuotient(int num1, int num2);
 }
